Added simple base workflow and changed some verbiage
